requirements.txt
aiosqlite==0.13.0
click==7.1.2
fastapi==0.56.0
h11==0.9.0
httptools==0.1.1
iso8601==0.1.12
pydantic==1.5.1
PyPika==0.37.7
starlette==0.13.2
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
